refactor/backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
werkzeug
pycryptodome
pymysql
cryptography
flask-mail
yfinance
pandas
numpy
google-generativeai
python-dotenv
python-dateutil
requests
akshare
alpha_vantage
apscheduler
stripe
supabase
psycopg2-binary

